bluetooth by device name, adapt restaurant name dynamically

+++ app/src/main/java/de/uni/hannover/hci/mi/team6/covidcheckin/model/Address.kt
@@ -3,4 +3,4 @@ package de.uni.hannover.hci.mi.team6.covidcheckin.model
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Address(val street: String, val streetNumber: String, val zipCode: Int, val city: String)
+data class Address(val street: String, val streetNumber: String, val zipCode: Long, val city: String)
